\name{mapping}
\alias{mapping}
\alias{mapping}
\title{A mapping of Refseq Protein IDs to probe set IDs for the gene expression data...}
\description{A mapping of Refseq Protein IDs to probe set IDs for the gene expression data}
\details{A mapping of the hgu95av2 probe set IDs in x to the Refseq protein IDs contained
in adjacency.matrix}
\docType{data}
\author{Marc Johannes \email{M.Johannes@DKFZ.de}}
\keyword{data}

