\docType{package}
\name{pathClass-package}
\alias{pathClass}
\alias{pathClass-package}
\title{Classification with SMVs and prior knowledge}
\description{
  Classification with SMVs and prior knowledge
}
\details{
  pathClass is a collection of classification methods that
  use information about how features are connected in the
  underlying biological network as an additional source of
  information. This additional knowledge is incorporated
  into the classification a priori. Several authors have
  shown that this approach significantly increases the
  classification performance.
}
\author{
  Marc Johannes \email{JohannesMarc@gmail.com}
}
\keyword{package}

