\name{y}
\alias{y}
\alias{y}
\title{Example class labels for the gene expression data...}
\description{Example class labels for the gene expression data}
\details{Class labels for the 10 patients contained the data matrix \code{x}}
\docType{data}
\author{Marc Johannes \email{JohannesMarc@gmail.com}}
\keyword{data}

