\name{adjacency.matrix}
\alias{adjacency.matrix}
\alias{adjacency.matrix}
\title{An adjacency matrix of a random graph...}
\description{An adjacency matrix of a random graph}
\details{An adjacency matrix of a random graph with some random Refseq
Protein IDs for use in example files and the vignette}
\docType{data}
\author{Marc Johannes \email{JohannesMarc@gmail.com}}
\keyword{data}

