\docType{data}
\name{mapping}
\alias{mapping}
\title{A mapping of Refseq Protein IDs to probe set IDs for the gene expression data}
\description{
  A mapping of the hgu95av2 probe set IDs in x to the
  Refseq protein IDs contained in adjacency.matrix
}
\author{
  Marc Johannes \email{JohannesMarc@gmail.com}
}
\keyword{data}
\keyword{datasets}

